alikson/teleport/my_version/pyproject.toml
[tool.poetry]
name = "my_version"
version = "0.1.0"
description = ""
authors = ["vadikam100500 <vadik.msk.100500@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Twisted = "21.2.0"
Pyrogram = "0.17.1"
xlrd = "^2.0.1"
xlwt = "^1.3.0"
demoji = "0.4.0"
pytz = "2021.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-broken-line = "^0.4.0"
flake8-return = "^1.1.3"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
